+++ upcloud/service/storage.go
@@ -31,6 +31,7 @@ type Storage interface {
 	GetStorageImportDetails(r *request.GetStorageImportDetailsRequest) (*upcloud.StorageImportDetails, error)
 	WaitForStorageImportCompletion(r *request.WaitForStorageImportCompletionRequest) (*upcloud.StorageImportDetails, error)
 	DeleteStorage(*request.DeleteStorageRequest) error
+	ResizeStorage(r *request.ResizeStorageRequest) (*upcloud.StorageResizeBackup, error)
 }
 
 var _ Storage = (*Service)(nil)
@@ -427,4 +428,28 @@ func (s *Service) WaitForStorageImportCompletion(r *request.WaitForStorageImport
 			time.Sleep(sleepDuration)
 		}
 	}
+}
+
+// ResizeStorage resizes the last partition of a storage and the ext3/ext4/XFS/NTFS filesystem
+// on that partition if the partition does not extend to the end of the storage yet.
+//
+// Before the resize is attempted, a backup is taken from the storage. If the resize
+// succeeds, backup details are returned. It is advisable to keep the backup until
+// you have ensured that everything works after the resize.
+//
+// If the resize fails, backup is used to restore the storage to the state where it
+// was before the resize. After that the backup is deleted automatically.
+func (s *Service) ResizeStorage(r *request.ResizeStorageRequest) (*upcloud.StorageResizeBackup, error) {
+	response, err := s.client.PerformJSONPostRequest(s.client.CreateRequestURL(r.RequestURL()), nil)
+
+	if err != nil {
+		return nil, parseJSONServiceError(err)
+	}
+
+	resizeBackup := upcloud.StorageResizeBackup{}
+	if err = json.Unmarshal(response, &resizeBackup); err != nil {
+		return &resizeBackup, nil
+	}
+
+	return &resizeBackup, nil
 }