+++ upcloud/storage.go
@@ -223,3 +223,33 @@ func (s *StorageImportDetails) UnmarshalJSON(b []byte) error {
 
 	return nil
 }
+
+// StorageResizeBackup represents the backup details taken before an resize operation.
+type StorageResizeBackup struct {
+	Access  string          `json:"access"`
+	Created time.Time       `json:"created"`
+	License float64         `json:"license"`
+	Origin  string          `json:"origin"`
+	Servers ServerUUIDSlice `json:"servers"`
+	Size    int             `json:"size"`
+	State   string          `json:"state"`
+	Title   string          `json:"title"`
+	Type    string          `json:"type"`
+	UUID    string          `json:"uuid"`
+	Zone    string          `json:"zone"`
+}
+
+// UnmarshalJSON is a custom unmarshaller that deals with deeply embedded values.
+func (s *StorageResizeBackup) UnmarshalJSON(b []byte) error {
+	type resizeBackup StorageResizeBackup
+	v := struct {
+		ResizeBackup resizeBackup `json:"resize_backup"`
+	}{}
+
+	if err := json.Unmarshal(b, &v); err != nil {
+		return err
+	}
+
+	*s = StorageResizeBackup(v.ResizeBackup)
+	return nil
+}